+++ server/requirements.txt
@@ -1,0 +1,110 @@
+annotated-types==0.7.0
+anyio==4.6.2.post1
+attrs==24.2.0
+black==24.10.0
+blinker==1.9.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+click==8.1.7
+contourpy==1.3.1
+coverage==7.6.7
+cycler==0.12.1
+cyclopts==2.9.9
+distro==1.9.0
+docstring_parser==0.16
+docutils==0.21.2
+easyocr==1.7.2
+filelock==3.16.1
+flake8==7.1.1
+Flask==3.1.0
+fonttools==4.55.0
+fsspec==2024.10.0
+gunicorn==23.0.0
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.27.2
+idna==3.10
+imageio==2.36.0
+iniconfig==2.0.0
+itsdangerous==2.2.0
+Jinja2==3.1.4
+jiter==0.7.1
+kiwisolver==1.4.7
+lazy_loader==0.4
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+matplotlib==3.9.2
+mccabe==0.7.0
+mdurl==0.1.2
+mpmath==1.3.0
+mypy==1.13.0
+mypy-extensions==1.0.0
+networkx==3.4.2
+ninja==1.11.1.1
+numpy==2.1.3
+nvidia-cublas-cu12==12.4.5.8
+nvidia-cuda-cupti-cu12==12.4.127
+nvidia-cuda-nvrtc-cu12==12.4.127
+nvidia-cuda-runtime-cu12==12.4.127
+nvidia-cudnn-cu12==9.1.0.70
+nvidia-cufft-cu12==11.2.1.3
+nvidia-curand-cu12==10.3.5.147
+nvidia-cusolver-cu12==11.6.1.9
+nvidia-cusparse-cu12==12.3.1.170
+nvidia-nccl-cu12==2.21.5
+nvidia-nvjitlink-cu12==12.4.127
+nvidia-nvtx-cu12==12.4.127
+openai==1.54.4
+opencv-contrib-python==4.10.0.84
+opencv-python==4.10.0.84
+opencv-python-headless==4.10.0.84
+packaging==24.2
+pandas==2.2.3
+pastel==0.2.1
+pathspec==0.12.1
+pillow==11.0.0
+platformdirs==4.3.6
+pluggy==1.5.0
+poethepoet==0.29.0
+psutil==6.1.0
+py-cpuinfo==9.0.0
+pyclipper==1.3.0.post6
+pycodestyle==2.12.1
+pydantic==2.9.2
+pydantic_core==2.23.4
+pyflakes==3.2.0
+Pygments==2.18.0
+pyparsing==3.2.0
+pytest==8.3.3
+pytest-cov==5.0.0
+pytest-socket==0.7.0
+python-bidi==0.6.3
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+pytz==2024.2
+PyYAML==6.0.2
+requests==2.32.3
+requests-mock==1.12.1
+rich==13.9.4
+rich-rst==1.3.1
+scikit-image==0.24.0
+scipy==1.14.1
+seaborn==0.13.2
+setuptools==75.5.0
+shapely==2.0.6
+six==1.16.0
+sniffio==1.3.1
+sympy==1.13.1
+-e git+https://github.com/David1000734/GROUP1-Tabby.git@6bbebfeebf7b757f83d5a81fd5ad8b38e2691ccc#egg=tabby_server&subdirectory=server
+tifffile==2024.9.20
+torch==2.5.1
+torchvision==0.20.1
+tqdm==4.67.0
+triton==3.1.0
+typing_extensions==4.12.2
+tzdata==2024.2
+ultralytics==8.3.33
+ultralytics-thop==2.0.11
+urllib3==2.2.3
+waitress==3.0.2
+Werkzeug==3.1.3
