+++ server/tabby_server/__main__.py
@@ -1,48 +1,12 @@
-from flask import Flask
-from http import HTTPStatus
-from .services import library
-from .vision import yolo_test
-from .api import books
-
-"""
-This is the central file of our app. Everything is called from here.
-
-Blueprints will lead to other files and packages so they can be accessed.
-
-The functions here are merely tests and should not actually be used.
-"""
-
-app = Flask(__name__)
-
-# Temporary Python Files. These are temporary routing to enable testing
-# of functionality. In production, these routing would be removed or unused.
-
-# Testing Python Files.
-# Google Books Implementation - 1 Routable Function (search)
-app.register_blueprint(library.books_api, url_prefix="/library")
-# You Only Look Once (YOLO) Implementation - 1 Routable Function (shelf_read)
-app.register_blueprint(yolo_test.yolo_test, url_prefix="/yolo")
-# Front End Endpoints - 2 Routable Functions (scan_cover and search)
-app.register_blueprint(books.subapp, url_prefix="/books")
+from tabby_server import app
+import os
 
 
-# Blueprints documentation
-# https://flask.palletsprojects.com/en/stable/blueprints/
-
-
-# Members API route
-@app.route("/members", methods=["GET"])
-def members():
-    return {"members": ["Member1", "Member2", "Member3"]}, HTTPStatus.OK
-
-
-@app.route("/api/test", methods=["POST"])
-def test():
-    return {"message": "Hello world!"}, HTTPStatus.OK
+PORT = os.getenv("PORT", "8000")
 
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    app.run(debug=False, host='0.0.0.0', port=int(PORT))
 
 # Run by using
 # python3 server.py
